+++ src/Parser/Php/PhpClassReference.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace Symfony\Lsp\Parser\Php;
+
+final class PhpClassReference
+{
+    public function __construct(
+        private readonly string $className,
+        private readonly int $startOffset,
+        private readonly int $endOffset,
+    ) {
+    }
+
+    public function className(): string
+    {
+        return $this->className;
+    }
+
+    public function startOffset(): int
+    {
+        return $this->startOffset;
+    }
+
+    public function endOffset(): int
+    {
+        return $this->endOffset;
+    }
+}